Rename package auth -> googleapi
